fix wrong names for universal objects
