+++ lib/mindee/parsing/common.rb
@@ -1,12 +1,15 @@
 # frozen_string_literal: true
 
+require_relative 'common/api_request'
 require_relative 'common/api_response'
 require_relative 'common/document'
 require_relative 'common/execution'
 require_relative 'common/execution_file'
 require_relative 'common/execution_priority'
 require_relative 'common/inference'
+require_relative 'common/job'
 require_relative 'common/ocr'
 require_relative 'common/prediction'
 require_relative 'common/orientation'
 require_relative 'common/page'
+require_relative 'common/workflow_response'
